fmt_to_flags: use JSON_PRESERVE_ORDER to fix unit test
The 'objects' unit test for md_json fails nondeterministically for
versions of Jansson < 2.8, because the order of name/value pairs isn't
defined by default, and the serialization tests do a straight string
comparison. To fix it, pass JSON_PRESERVE_ORDER in the dump flags (this
is now the default in Jansson 2.8, and there is no way to turn it off).