Send "Content-Length" only, without "Transfer-Encoding: chunked"
libcurl 7.66.0 automatically adds a "Transfer-Encoding: chunked" header
if the content length is not known to libcurl, because of this change:
https://github.com/curl/curl/pull/4138

libcurl does not parse custom "Content-Length" headers, it needs the
content length using the options CURLOPT_POSTFIELDSIZE(_LARGE) and
CURLOPT_INFILESIZE(_LARGE).

The "Boulder" server rejects requests with both headers, but the host
acme-v02.api.letsencrypt.org accepts such requests, probably because
there's a reverse proxy.