Improve protocol check for pre-configured "tls-alpn-01" challenge
For "tls-alpn-01" challenges, mod_md checks whether there is a server
that supports the "acme-tls/1" protocol.

If the challenge "tls-alpn-01" is pre-configured with MDCAChallenges,
the ports of the servers are ignored (see issue #133). If there are
multiple servers matching a domain name, don't just look at the first
server but search one that supports the "acme-tls/1" protocol.

This is a follow-up to https://github.com/icing/mod_md/pull/159 and
https://github.com/icing/mod_md/issues/133