Handle extracting iPAddresses in SAN correctly
IPAddresses are encoded as ASN1_OCTECT_STRINGs, not ia5.

They have the odd property that IPv4 and IPv6 are not labeled,
but are distinguished by length.

Convert to text as the code intended. (No compression of IPv6).

Note: LE plans to support IP addresses in their certificates,
though it's not clear how mod_md should handle.  Perhaps it will
be a MDNameTypes DNS IPADDRESS directive to setup the CSR with
either or both.

This commit only corrects extracting IP addresses from a certificate.

More work is required...

Ref: https://letsencrypt.org/upcoming-features/
     https://tools.ietf.org/html/rfc5280#section-4.2.1.6