Add contrib directory.  Initial SeLinux policy tweaks.
The SeLinux policy tweaks are those added when I enabled
mod_md.  They may not suffice on an unmodified system,
since I had other tweaks in place for httpd.  However, it's a
good start.

Note that the usual iterative process of finding/fixing the
necessary changes can have intermediate states that confuse
mod_md.

The intent of contrib is that other tools can be added.  For example,
a more functional md_message.sh, programs that consume the json
status, etc.