Preliminary version of contrib/md_events
This script is a preliminary version of a handler for the
md_events discussed in issue #189.  It has gone through some
unit testing, but is not production qualified.  Nothing about
it should be considered stable.

This version is to drive discussion of the mod_md interfaces.

It should work as an MDMessageCmd and MDChallengeDns01 provider,
as well as for the (hypothetical) MDChallengeHttp01 events.

Feedback and testing is encouraged.

To get started, run contrib/md_events/md_events -h for usage.

Then contrib/md_events/md_events -I, which will do a default install.

/etc/httpd/md_events.conf and /etc/httpd/md_events.d will be
informative.

You can play with it from the command line as installed:
  contrib/md_events/md_events -uroot installed example.net

I have also provided a few dns-01 update scripts; md_events has
an (untested) wrapper that should make them useful.