md_events - optimize install, work-around dns-01 tokens
When installing, remove the installation code.  Besides making
the script smaller, the installed version is easier to read without
all the template configuration files...

The existing DNS update scripts expect the token to be deleted.
mod_md doesn't provide it.  For now, keep track of tokens that
are installed, and remove them (all) using the index.  This is
ugly - but should work and allow the library of scripts to work
unmodified.