Fix lost environment variables
Adding the environment variables for MDMessageCmd and friends had
the unintended side-effect of passing ONLY those variables (and
a couple of others) to the child.  This means no Path on Windows,
among other things.

#198 reported this issue.

This patch should work - except that technically, the environment
variables aren't thread-safe.  The right thing is probably to add
a mutex - but figuring out how APR spells that is for another day.  @icing?