Correctly report ACME problem details
When the "application/problem+json" header includes parameters
(e.g. ";charset=utf-8", inspect_problem misclassified the response.

Only compare the initial part of the content-type.

It would be more correct to look for the optional ';', but
the same approach is used elsewhere in the code...they can
all be cleaned up at some future point.